Consolidate TAXYEAR imports to canonical source
All 9 files that imported TAXYEAR from
tmd.create_taxcalc_input_variables (a re-export) now import
directly from tmd.imputation_assumptions (the single source
of truth where TAXYEAR is defined).
