+++ tests/test_misc.py
@@ -5,7 +5,7 @@ Miscellaneous tests of tmd.csv variable weighted totals.
 import pytest
 import taxcalc as tc
 from tmd.storage import STORAGE_FOLDER
-from tmd.create_taxcalc_input_variables import TAXYEAR
+from tmd.imputation_assumptions import TAXYEAR
 from tests.conftest import create_tmd_records
