+++ tests/test_tax_revenue.py
@@ -5,7 +5,7 @@ Tests of income/payroll tax revenues generated by the tmd files.
 import yaml
 import numpy as np
 import taxcalc as tc
-from tmd.create_taxcalc_input_variables import TAXYEAR
+from tmd.imputation_assumptions import TAXYEAR
 
 FIRST_CYR = 2023
 LAST_CYR = 2033
