+++ tmd/create_taxcalc_sampling_weights.py
@@ -7,7 +7,7 @@ create_taxcalc_input_variable.py script.
 import yaml
 import pandas as pd
 from tmd.storage import STORAGE_FOLDER
-from tmd.create_taxcalc_input_variables import TAXYEAR
+from tmd.imputation_assumptions import TAXYEAR
 
 FIRST_YEAR = TAXYEAR
 LAST_YEAR = 2074
